+++ src/collective/weather/weather_utility.py
@@ -1,17 +1,16 @@
 # -*- coding: utf-8 -*-
 
-from collective.weather.interfaces import IWeatherSettings
 from collective.weather.config import COOKIE_KEY
 from collective.weather.config import PROJECTNAME
 from collective.weather.config import TIME_THRESHOLD
 from collective.weather.interfaces import IWeatherInfo
+from collective.weather.interfaces import IWeatherSettings
 from collective.weather.interfaces import IWeatherUtility
 from datetime import datetime
+from plone import api
 from plone.registry.interfaces import IRegistry
-from Products.CMFCore.utils import getToolByName
 from zope.component import getUtility
 from zope.component import queryUtility
-from zope.component.hooks import getSite
 from zope.globalrequest import getRequest
 from zope.interface import implements
 
@@ -74,8 +73,7 @@ class WeatherUtility(object):
             utility = queryUtility(IWeatherInfo, name=provider)
             if utility:
                 weather_api = utility(api_key)
-                portal = getSite()
-                ltool = getToolByName(portal, 'portal_languages')
+                ltool = api.portal.get_tool('portal_languages')
                 try:
                     result = weather_api.getWeatherInfo(cityid, units=units, lang=ltool.getDefaultLanguage())
                 except Exception, msg: